internal/cabi: support arm

+++ _demo/cabi/wrap/wrap.c
@@ -127,31 +127,31 @@ struct point7 pt7(struct point7 pt) {
 
 struct data1 {
     char x;
-    size_t y;
+    long long y;
 };
 
 struct data1 fn1(struct data1 pt) {
-    printf("data1: %d %zu\n",pt.x,pt.y);
+    printf("data1: %d %lld\n",pt.x,pt.y);
     return pt;
 }
 
 struct data2 {
     int x;
-    size_t y;
+    long long y;
 };
 
 struct data2 fn2(struct data2 pt) {
-    printf("data2: %d %zu\n",pt.x,pt.y);
+    printf("data2: %d %lld\n",pt.x,pt.y);
     return pt;
 }
 
 struct data3 {
-    size_t x;
+    long long x;
     char y;
 };
 
 struct data3 fn3(struct data3 pt) {
-    printf("data3: %zu %d\n",pt.x,pt.y);
+    printf("data3: %lld %d\n",pt.x,pt.y);
     return pt;
 }
